Added scripts for header
The header was (and still is) kind of boring, to improve user experience
of using the header navigation, I added support for:

Hovering a page link:
Some 'easy on the eye' hovering css on page links

Nested Pages
Pages with children will, in addition to getting a css hover-effect,
animate its width and display it's children in a new menu. This has not
been tested much.